Checking filetype before uploading + error handling

+++ server/srcs/app/controllers/Medias.controller.js
@@ -1,11 +1,13 @@
 const fs = require("fs");
 const uuid = require("uuid/v4");
+const {Magic, MAGIC_MIME_TYPE} = require("mmmagic");
 const MediasModel = require("../models/Medias.model");
 
 class MediasController {
     constructor(dtb) {
         this.medias = new MediasModel(dtb);
         this.mediaDir = "./srcs/imgs/";
+        this.validTypes = ["jpg", "jpeg", "png", "gif", "mp4"];
     }
     
     getName() {
@@ -19,14 +21,34 @@ class MediasController {
     }
     
     uploadFile(name, media, author) {
-        const filepath = `${this.mediaDir}${this.getName()}.${media.type.split("/")[1]}`;
-        try {
-            fs.writeFileSync(filepath, fs.readFileSync(media.path));
-            fs.unlinkSync(media.path);
-        } catch (err) {
-            return new Promise((resolve, reject) => reject(err));
-        }
-        return this.medias.addFile(name, filepath, author);
+        const magic = new Magic(MAGIC_MIME_TYPE);
+        return new Promise((resolve, reject) => {
+            magic.detectFile(media.path, (err, type) => {
+                if (err !== null) {
+                    return reject({statusCode: 500, message: err});
+                }
+                let match = false;
+                const extension = type.split("/")[1];
+                this.validTypes.map(validType => {
+                    if (extension === validType) {
+                        match = true;
+                    }
+                });
+                if (!match) {
+                    return reject({statusCode: 400, message: "Filetype is not valid"});
+                }
+                const filepath = `${this.mediaDir}${this.getName()}.${extension}`;
+                try {
+                    fs.writeFileSync(filepath, fs.readFileSync(media.path));
+                    fs.unlinkSync(media.path);
+                } catch (err) {
+                    return reject({statusCode: 500, message: err});
+                }
+                this.medias.addFile(name, filepath, author, type)
+                    .then(result => resolve(result))
+                    .catch(err => reject({statusCode: 500, message: err}));
+            });
+        });
     }
 
     getAll(page, limit) {
